pyproject.toml
[tool.poetry]
name = "country-monitor"
version = "1.0.0"
description = "Country Monitor App"
authors = ["amartilotta <agustin.martilotta@hotmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
django = "^5.1.4"
djangorestframework = "^3.15.2"
redis = "^5.2.1"
celery = "^5.4.0"
flower = "^2.0.1"
requests = "^2.32.3"
psycopg2-binary = "^2.9.10"
django-filter = "^24.3"


[tool.poetry.group.dev.dependencies]
mypy = "^1.10.1"
black = "^24.4.2"
ruff = "^0.5.3"
ipdb = "^0.13.13"
debugpy = "^1.8.2"
pickleshare = "^0.7.5"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
pytest-sugar = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
mypy_path = ["src"]
plugins = [
  "pydantic.mypy"
]
ignore_missing_imports = true
explicit_package_bases = true
check_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.coverage.run]
relative_files = true

[tool.ruff]
line-length = 79
exclude = [".env", ".venv", "venv"]
output-format = "concise"

[tool.ruff.lint]
select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    "W",
    # isort
    "I001"
]
ignore = ["E501"]
fixable = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]

[tool.black]
line-length = 79
force-exclude = '''
/(
  | .env
  | .venv
  | .mako
  | venv
  | notebooks
)/
'''